% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{add_pseudo_obs}
\alias{add_pseudo_obs}
\title{Add pseudo observations to a data frame}
\usage{
add_pseudo_obs(data, tme = c(17.5, 20, 26.5, 35))
}
\arguments{
\item{data}{A data frame with censored survival times for 2 outcomes}

\item{tme}{Vector of times at which to compute pseudo values}
}
\value{
The data frame with pseudo observations added for t = 26.5 weeks
}
\description{
Allow arbitrary names
}
