% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_roc}
\alias{calc_roc}
\title{Compute the time-varying cause-specific ROC curve using pseudo-observations}
\usage{
calc_roc(X, pseu, pseu2)
}
\arguments{
\item{X}{Vector of predictions}

\item{pseu}{Pseudo observations at a fixed time for the cause of interest}

\item{pseu2}{Pseudo observations for the other cause}
}
\description{
Compute the time-varying cause-specific ROC curve using pseudo-observations
}
