% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate survival data for analysis}
\usage{
generate_data()
}
\value{
A data frame with X variables, censored survival times (competing risk w death), and true cumulative incidence at 26.5 weeks
}
\description{
500 subjects, 20 variables, about 20% cumulative incidence. 5 variables truly associated,
some linear, some wonky associations, an interaction, and some variables correlated.
}
