% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{local_slope}
\alias{local_slope}
\title{Perturbation algorithm to get local slopes}
\usage{
local_slope(fitted, data, predictors)
}
\arguments{
\item{fitted}{The fitted model object, e.g. from superlearner_estimate}

\item{data}{A dataframe to be perturbed}

\item{predictors}{the predictors used the the model}
}
\value{
A data frame with the local slopes for each variable (columns) and each subject (rows)
}
\description{
Given a fitted model object with a predict method, run a perturbation algorithm
that estimates the local slope of the predicted outcome for each predictor in the
model and for each subject in the dataset.
}
