% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optimize_auc}
\alias{optimize_auc}
\title{Optimize the time varying AUC}
\usage{
optimize_auc(Z, Y, pseu2)
}
\arguments{
\item{Z}{Matrix of predictions}

\item{Y}{Pseudo values for the cumulative incidence for the competing event of interest at a fixed time}

\item{pseu2}{Matrix of pseudo values for the cumulative incidences for all other competing events}
}
\description{
Optimize the time varying AUC
}
