% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_generated_data.R
\name{pcox_estimate}
\alias{pcox_estimate}
\title{Run the a penalized cox regression model}
\usage{
pcox_estimate(formula, data)
}
\arguments{
\item{formula}{A formula describing the model}

\item{data}{A data frame with a survival outcome and a set of predictors}
}
\value{
A fitted model object
}
\description{
Run the a penalized cox regression model
}
