% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.R
\name{pseudoci.weighted}
\alias{pseudoci.weighted}
\title{Pseudo observations for the weighted cumulative incidence function}
\usage{
pseudoci.weighted(time, event, tmax, weights = rep(1, length(time)))
}
\arguments{
\item{time}{follow up time}

\item{event}{cause indicator, use 0 as censoring code and integers for other causes}

\item{tmax}{a vector of time points at which pseudo-observations are to be computed. If missing, uses each event time}

\item{weights}{a vector of case weights, i.e., 1/probabilities}
}
\value{
A list containing the following objects
\describe{
  \item{time}{The ordered time points at which the pseudo-observations are evaluated.}
  \item{cause}{The ordered codes for different causes.}
  \item{pseudo}{A list of matrices - a matrix for each of the causes, ordered by codes. Each row of a matrix belongs to one individual (ordered as in the original data set), each column presents a time point (ordered in time).}
}
}
\description{
Computes pseudo-observations for modeling competing risks based on the possibly weighted cumulative incidence function
}
