% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{shift_pred}
\alias{shift_pred}
\title{Perturbation algorithm to get local slopes}
\usage{
shift_pred(SLobject, X, predictor)
}
\arguments{
\item{SLobject}{The fitted model object, e.g. from superlearner_estimate}

\item{X}{A dataframe to be perturbed}

\item{predictor}{the predictors used the the model}
}
\value{
A data frame with the local slopes for each variable (columns) and each subject (rows)
}
\description{
Given a fitted model object with a predict method, run a perturbation algorithm
that estimates the local slope of the predicted outcome for each predictor in the
model and for each subject in the dataset.
}
