% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_generated_data.R
\name{stupidlearner_estimate}
\alias{stupidlearner_estimate}
\title{For testing purposes}
\usage{
stupidlearner_estimate(data, Y = "cause1.pseudo", X = paste0("X",
  1:20), y0, y1)
}
\description{
For testing purposes
}
