% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_generated_data.R
\name{superlearner_binaryestimate}
\alias{superlearner_binaryestimate}
\title{Run the superlearner model with binary indicator as the outcome}
\usage{
superlearner_binaryestimate(data, Y = "binY", X = paste0("X", 1:20))
}
\arguments{
\item{data}{A data frame with pseudo observations and a set of predictors}

\item{Y}{character string for the outcome variable in data}

\item{X}{character vector for the predictor variables in data}
}
\value{
A fitted model object
}
\description{
Run the superlearner model with binary indicator as the outcome
}
