% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_generated_data.R
\name{superlearner_estimate}
\alias{superlearner_estimate}
\title{Run the superlearner model with pseudo observations as the outcome}
\usage{
superlearner_estimate(formula, data)
}
\arguments{
\item{formula}{A formula describing the model}

\item{data}{A data frame with pseudo observations and a set of predictors}
}
\value{
A fitted model object
}
\description{
Run the superlearner model with pseudo observations as the outcome
}
