% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate survival data for analysis}
\usage{
generate_data(n = 500, scenario = "A", missing.p = 0.1,
  missing.site = FALSE)
}
\arguments{
\item{n}{Sample size}

\item{scenario}{Character designating the simulation scenario: 0 = NULL, no association, A = simple, one variable
is linearly associated with IRIS,
B = simple multivariate, 5 variables are linearly associated,
C = wonky, interactions, nonlinearities, etc.}

\item{missing.p}{Proportion of missing binary Y values (random censoring)}

\item{missing.site}{logical, if TRUE, censoring depends on a binary variable}
}
\value{
A data frame with X variables, censored survival times (competing risk w death), and true cumulative incidence at 26.5 weeks
}
\description{
500 subjects, 20 variables, about 20% cumulative incidence. 5 variables truly associated,
some linear, some wonky associations, an interaction, and some variables correlated.
}
